% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampEn_C.R
\name{SampEn}
\alias{SampEn}
\title{SampEn}
\usage{
SampEn(TS, dim = 2, lag = 1, r = 0.2 * sd(TS))
}
\arguments{
\item{TS}{- given time series}

\item{dim}{- dimension of given time series, default value is 2}

\item{lag}{- downsampling, default value is 1}

\item{r}{- radius of searched areas, default value is 0.2*sd(TS)}
}
\description{
This function computes sample entropy of given time series.
}
\examples{
timser <- rnorm(2000)
SampEn(timser)
SampEn(timser, r = 0.1*sd(timser))
SampEn(timser, dim = 3, r = 0.1*sd(timser))

}
\keyword{SampEn}
\keyword{entropy}
\keyword{sample}
